[#3748] feat: move types import
