+++ src/bentoml/_internal/service/openapi/utils.py
@@ -5,42 +5,18 @@ import typing as t
 from typing import TYPE_CHECKING
 from functools import lru_cache
 
-if sys.version_info >= (3, 8):
-    from typing import get_args
-    from typing import TypedDict
-    from typing import get_origin
-else:
-    from typing_extensions import get_args
-    from typing_extensions import TypedDict
-    from typing_extensions import get_origin
-
-if sys.version_info >= (3, 10):
-    from types import UnionType
-    from typing import is_typeddict
-    from typing import get_type_hints
-
-    def is_union(tp):  # check Union[A,B] or A | B
-        return tp == t.Union or get_origin(tp) == UnionType
-
-else:
-    from typing_extensions import is_typeddict
-    from typing_extensions import get_type_hints
-
-    def is_union(tp):  # check Union[A,B]
-        return tp == t.Union
-
-
-if sys.version_info >= (3, 11):
-    from typing import Required
-    from typing import NotRequired
-else:
-    from typing_extensions import Required
-    from typing_extensions import NotRequired
-
 from bentoml.exceptions import NotFound
 from bentoml.exceptions import InvalidArgument
 from bentoml.exceptions import InternalServerError
 
+from ...types import get_args
+from ...types import is_union
+from ...types import Required
+from ...types import TypedDict
+from ...types import get_origin
+from ...types import NotRequired
+from ...types import is_typeddict
+from ...types import get_type_hints
 from ...utils import LazyLoader
 from .specification import Schema
